building out the skeleton
