+++ src/portals/MousePortal/components/mouse-footer.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    Mouse footer coming soon
+  </div>
+</template>
