+++ src/portals/MousePortal/components/mouse-header.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    Mouse header coming soon
+  </div>
+</template>
