+++ src/portals/MousePortal/views/Contact/Template.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    Contact Us!
+  </div>
+</template>
