+++ src/portals/MousePortal/views/Index/Template.vue
@@ -1,0 +1,3 @@
+<template>
+  Hello Mouse World!
+</template>
